Modification of `ph_bw_update_lambda'.
This function does not update the lambda parameters; it only computes a
possible update. Whether this update is apply or not is decided after
the M-Step. So, this commit renames `ph_bw_update_lambda' to
`ph_bw_m_step_lambda'.

The state probabilities , computed by the call to `mm_add_s' are also
needed later when the delta update is computed. Hence, this call must be
placed outside of `ph_bw_m_step_lambda'. This commit just removes the
call.

Also, this commit chages the interfaces, which is in part a side effect
of removing `mm_add_s'. In addition to that, rename the output buffer
`lambda_update' to `out'.