Log conditional expectation.
Declare and define `log_cond_expect', a low-level function to
compute the logarithm of the "conditional expectations of those function
of the  missing data that appear in the complete-data log-likelihood"
(see Zucchini, p. 64).

Define the parameterized macro `PoisHmm_LogConditionalExpectation',
which provides an high-level interface to `log_cond_expect'.

Remove `PoisHmm_LogStateProbs', which implemented the same
functionallity as `log_cond_expect'. However, the name did not reflect
this.