Rewrite sentinel
Python documentation explains to implement the sentinel with `{NULL}`.
However, this triggers a -Wunused-filed warning in clang.

This implementation fills in ever field of the struct to get rid of the
warning. The -1 in the type field (second field) ensures that no define
Python type is addressed by accident.

The types are defined here:
https://github.com/python/cpython/blob/main/Include/structmember.h